+++ source/constants.h
@@ -10,5 +10,8 @@ PyObject *rising_edge;
 PyObject *falling_edge;
 PyObject *both_edge;
 PyObject *version;
+PyObject *unknown;
+PyObject *board;
+PyObject *bcm;
 
 void define_constants(PyObject *module);